refactor: convert map details panel from innerHTML to React component
